Refactored package name to com.emergenseek.client
