+++ lib/main.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/screens/home.dart';
-import 'package:flutter_app/screens/service_locator.dart';
-import 'package:flutter_app/screens/location_updates.dart';
-import 'package:flutter_app/screens/sos.dart';
-import 'package:flutter_app/screens/profile.dart';
-import 'package:flutter_app/screens/contacts.dart';
-import 'package:flutter_app/screens/about.dart';
-import 'package:flutter_app/models/sos_model.dart';
+import 'package:EmergenSeek/screens/home.dart';
+import 'package:EmergenSeek/screens/service_locator.dart';
+import 'package:EmergenSeek/screens/location_updates.dart';
+import 'package:EmergenSeek/screens/sos.dart';
+import 'package:EmergenSeek/screens/profile.dart';
+import 'package:EmergenSeek/screens/contacts.dart';
+import 'package:EmergenSeek/screens/about.dart';
+import 'package:EmergenSeek/models/sos_model.dart';
 import 'package:scoped_model/scoped_model.dart';
 
 void main() {
