+++ lib/screens/about.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/screens/nav_menu.dart';
-import 'package:flutter_app/screens/sos_quick_button.dart';
+import 'package:EmergenSeek/screens/nav_menu.dart';
+import 'package:EmergenSeek/screens/sos_quick_button.dart';
 
 class AboutPage extends StatelessWidget {
   @override
