+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: flutter_app
-description: A new Flutter application.
+name: EmergenSeek
+description: The Flutter client for the EmergenSeek platform.
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
